# doc/api/booking-core-p0.openapi.yaml
openapi: 3.0.3
info:
  title: booking-core P0 API (normative contract)
  version: "0.0.1"
  description: |
    Wave 1 (P0) **as implemented** in booking-core backend. Paths match running controllers under
    `/api/auth`, `/api/client`, `/api/merchant`.

    **Spec delta vs PM table (§4):** Login uses **username** + **password** (not email). Client
    “resource search/detail” for storefront is exposed as `GET /api/client/merchant/{slug}` (embeds
    services + resources). Merchant booking actions use a single
    `PUT /api/merchant/{merchantId}/bookings/{bookingId}/status` with `BookingStatus` body, not
    separate `/confirm` `/reject` `/complete` paths.

    **Error envelope** (ApiException / validation): JSON object `code` (string) and `message` (string).
servers:
  - url: /
...
components:
  securitySchemes:
    bearer-jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Value from `accessToken` after login.
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: INVALID_CREDENTIALS
        message:
          type: string
          example: Invalid credentials
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresInSeconds:
          type: integer
          format: int64
        role:
          type: string
        roles:
          type: array
          items:
            type: string
        canonicalRole:
          type: string
        canonicalRoles:
          type: array
          items:
            type: string
        roleAliases:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
...
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login (JWT access token)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              ok:
                summary: Valid body
                value:
                  username: client-demo
                  password: "secret123"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              examples:
                success:
                  value:
                    accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    tokenType: Bearer
                    expiresInSeconds: 3600
                    role: CLIENT
                    roles: [CLIENT]
                    permissions: []
        "400":
          description: Validation / bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                validation:
                  value:
                    code: VALIDATION_FAILED
                    message: password must not be blank
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                badCreds:
                  value:
                    code: API_BAD_REQUEST
                    message: Invalid credentials
        "429":
          description: Login rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                limited:
                  value:
                    code: API_BAD_REQUEST
                    message: Too many login attempts
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current principal (roles, permissions, contexts)
      security:
        - bearer-jwt: []
      responses:
        "200":
          description: Profile snapshot
          content:
            application/json:
              schema:
                type: object
                description: AuthMeResponse — see runtime `/v3/api-docs` for full record fields
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
              examples:
                unauthorized:
                  value:
                    code: API_BAD_REQUEST
                    message: Unauthorized
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (stateless no-op; client discards token)
      security:
        - bearer-jwt: []
      responses:
        "204":
          description: No content
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...